% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcDyspiaStatCumulative}
\alias{calcDyspiaStatCumulative}
\title{Calculates DysPIA statistic values for all the prefixes of a gene pair set}
\usage{
calcDyspiaStatCumulative(stats, selectedStats, DyspiaParam)
}
\arguments{
\item{stats}{Named numeric vector with gene pair-level statistics
sorted in decreasing order (order is not checked)}

\item{selectedStats}{indexes of selected gene pairs in a `stats` array}

\item{DyspiaParam}{DysPIA weight parameter (0 is unweighted, suggested value is 1)}
}
\value{
Numeric vector of DysPIA statistics for all prefixes of selectedStats.
}
\description{
Calculates DysPIA statistic values for all the prefixes of a gene pair set
}
