% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcDyspiaStatCumulativeBatch}
\alias{calcDyspiaStatCumulativeBatch}
\title{Calculates DysPIA statistic values for the gene pair sets}
\usage{
calcDyspiaStatCumulativeBatch(
  stats,
  DyspiaParam,
  pathwayScores,
  pathwaysSizes,
  iterations,
  seed
)
}
\arguments{
\item{stats}{Named numeric vector with gene pair-level statistics 
sorted in decreasing order (order is not checked).}

\item{DyspiaParam}{DysPIA weight parameter (0 is unweighted, suggested value is 1).}

\item{pathwayScores}{Vector with enrichment scores for the pathways in the database.}

\item{pathwaysSizes}{Vector of pathway sizes.}

\item{iterations}{Number of iterations.}

\item{seed}{Seed vector}
}
\value{
List of DysPIA statistics for gene pair sets.
}
\description{
Calculates DysPIA statistic values for the gene pair sets
}
