% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calEdgeCorScore_ESEA.R
\name{calEdgeCorScore_ESEA}
\alias{calEdgeCorScore_ESEA}
\title{calEdgeCorScore_ESE}
\usage{
calEdgeCorScore_ESEA(
  dataset,
  class.labels,
  controlcharacter,
  casecharacter,
  background
)
}
\arguments{
\item{dataset}{Matrix of gene expression values (rownames are genes, columnnames are samples).}

\item{class.labels}{Vector of binary labels.}

\item{controlcharacter}{Charactor of control in the class labels.}

\item{casecharacter}{Charactor of case in the class labels.}

\item{background}{Matrix of the edges' background.}
}
\value{
A vector of the aberrant correlation in phenotype P based on mutual information (MI) for each edge.
}
\description{
Calculates differential Mutual information.
}
\examples{
data(gene_expression_p53, class.labels_p53,sample_background)
ESEAscore_p53<-calEdgeCorScore_ESEA(gene_expression_p53, class.labels_p53,
 "WT", "MUT", sample_background)

}
