% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{class.labels_p53}
\alias{class.labels_p53}
\title{Example vector of category labels.}
\usage{
data(class.labels_p53)
}
\description{
The labels for the 50 cell lines in p53 data. Control group's label is 'WT', case group's label is 'MUT'.
}
