% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{DysGPS_p53}
\alias{DysGPS_p53}
\title{Example vector of DysGPS in p53 data.}
\usage{
data(DysGPS_p53)
}
\description{
The score vector of 164923 gene pairs from p53 dataset. 
It can be loaded from the example datasets of R-package 'DysPIA', 
and also can be obtained by running DysGPS(), details see DysGPS.R
}
