% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{DyspiaRes_p53}
\alias{DyspiaRes_p53}
\title{Example list of DysPIA result in p53 data.}
\usage{
data(DyspiaRes_p53)
}
\description{
The list includes 81 pathway results from 'DisPIA.R' as an example used in 'DyspiaSig.R'.
}
