% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyspiaSig.R
\name{DyspiaSig}
\alias{DyspiaSig}
\title{DyspiaSig}
\usage{
DyspiaSig(DyspiaRes, fdr)
}
\arguments{
\item{DyspiaRes}{Table with results of running DysPIA().}

\item{fdr}{Significant threshold of `padj` (a BH-adjusted p-value).}
}
\value{
A list of significant DysPIA results, including correlation gain and correlation loss.
}
\description{
Returns the significant summary of DysPIA results.
}
\examples{
data(pathway_list,package="DysPIAData")
data(DyspiaRes_p53)
summary_p53 <- DyspiaSig(DyspiaRes_p53, 0.05)       # filter with padj<0.05

}
