% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{gene_expression_p53}
\alias{gene_expression_p53}
\title{Example matrix of gene expression value.}
\usage{
data(gene_expression_p53)
}
\description{
A dataset of transcriptional profiles from p53+ and p53 mutant cancer cell lines. 
It includes the normalized gene expression for 6385 genes in 50 samples.
Rownames are genes, columnnames are samples.
}
