% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\name{sample_background}
\alias{sample_background}
\title{Example list of gene pair background.}
\usage{
data(sample_background)
}
\description{
The list of background was used in ''DysGPS.R' and 'calEdgeCorScore_ESEA.R' which is a part of the 'combined_background' in 'DysPIAData'.
}
