% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysPIA.R
\name{setUpBPPARAM}
\alias{setUpBPPARAM}
\title{setUpBPPARAM}
\usage{
setUpBPPARAM(nproc = 0, BPPARAM = NULL)
}
\arguments{
\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 0).}

\item{BPPARAM}{Parallelization parameter used in bplapply.
Can be used to specify cluster to run. If not initialized explicitly or
by setting `nproc` default value `bpparam()` is used.}
}
\value{
parameter BPPARAM value
}
\description{
Sets up parameter BPPARAM value.
}
